MATERIALS CHANGE: permit arrays with length = 1.
Please recompile your materials after this change.

This basically changes our sentinel for non-arrays from 1 to 0 in order
to allow for 1-sized array parameters, which is a feature that will make
life easier for our clients.

As a side effect, this fixes a behavioral mismatch between the
programmatic MaterialBuilder and matc, because the latter was already
using 0 as the sentinel. (look for arraySize in ParametersProcessor.cpp)

Fixes #753.